==> analysis_rmd/1_HTO_diagnostics.Rmd <==
---
title: "HTO diagnostics"
author: "Stephany Orjuela & Anthony Sonrel"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
    theme: yeti
    highlight: tango
    code_folding: show
    keep_md: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, dev = c("png", "pdf"))
```

## Load packages

```{r}
suppressPackageStartupMessages({
    library(fishpond)
    library(tximport)
    library(devtools)
    library(magrittr)
    library(reshape2)
    library(SingleCellExperiment)
    library(ggplot2)
 })
```

```{r, eval = FALSE, include = FALSE}
# for interactive session to test one specific sample
fileName="W8"
inputFolder=paste0("alevin/HTO/",fileName,"/alevin/quants_mat.gz")
metadataFile="metadata/metadata.csv"
```



## Read alevin output

We load the alevin generated count matrices using tximport package.

First look at wells 1 and 3 since they are myeloid and easy to separate
```{r}

hto.files <- inputFolder

file.exists(hto.files)

hto <- tximport(files = hto.files, type = "alevin")

#read metadata
sample_key <- read.table(metadataFile,
                       header = TRUE,
                       sep = ",",
                       stringsAsFactors = FALSE)

rownames(sample_key) <- sample_key$sample_id

y <- as.matrix(sample_key[,grep("A0", colnames(sample_key))])
well <- which(sample_key$well==as.integer(gsub("W","",fileName)))
y <- y[well,]
#colnames(y) <- gsub("A","",colnames(y))
y <- y[,colSums(y)>0]
y

## Create SCE
d <- SingleCellExperiment(list(counts=hto$counts))

```


## Filter by HTO counts
```{r}
#x <- rowSums(hashtags >= 1) >= 1
x <- colSums(counts(d)) >= 50 # Default to 50 
table(x)
d <- d[,x]

#Get HTO library sizes
df <- data.frame(Barcode=colnames(d), t(as.matrix(counts(d))))
df$sum_hashtags <- colSums(counts(d))

head(df)
```

## Some distributions

```{r}
# distribution of number of hashtags per cell
ggplot(df %>% dplyr::filter(sum_hashtags < 500),aes(x=sum_hashtags)) +
  geom_density(size=2.5)  
  #facet_wrap(~Sample, nrow=2, scales = "free_y")

df1 <- data.frame(t(as.matrix(counts(d))))
df1m <- melt(df1)

# tags_by_sample #%>% dplyr::filter(value<1000)
ggplot(df1m %>% dplyr::filter(value<1000), aes(x=variable,y=value)) +
  geom_boxplot() +
  #scale_y_log10() +
  theme(axis.text.x = element_text(angle = 90)) +
  xlab("Barcodes") + 
  ggtitle(paste0("Well ", fileName,", barcodes ", paste(colnames(y), collapse = "+") ))
```

## Plot HTO combinations

```{r, fig.width = 10, fig.height= 12}

hashtags_1 <- as.data.frame(log10(df1 + 1))
colnames(hashtags_1) <- gsub("^X", "A0", colnames(hashtags_1) )
custom_func <- function(data, mapping){
  ggplot(data = data, mapping = mapping) +
    geom_point(alpha = 0.3, size = 0.5) +
    geom_density_2d(color = "orange")
}

plt <- try(print(GGally::ggpairs(hashtags_1, upper = NULL,
                lower = list(continuous = custom_func))), silent = TRUE)
if (class(plt) == "try-error") {
  # replacement as it may return an error for some low-quality hashtags
  pairs(hashtags_1, lower.panel = NULL,
      panel = function(...) smoothScatter(..., nrpoints = 0, add = TRUE))
} 

# plot the first hashtag of each sample only
ggplot(hashtags_1, aes(x=hashtags_1[,colnames(y)[y[1,] == 1][1]],
                       y=hashtags_1[,colnames(y)[y[2,] == 1][1]])) +
   geom_point(alpha = 0.3, size = 0.5) +
    geom_density_2d(color = "orange") + 
  xlab(colnames(y)[y[1,] == 1][1]) + 
  ylab(colnames(y)[y[2,] == 1][1])
```


### Look at proportions
```{r, fig.width = 12, fig.height= 14}
hash_sums <- rowSums(df1)
hash_prop <- as.data.frame(apply(df1, 2, function(u) u/hash_sums))
colnames(hash_prop) <- gsub("^X", "A0", colnames(hash_prop) )
head(hash_prop)

plt <- try(print(GGally::ggpairs(hash_prop, upper = NULL,
                                 lower = list(continuous = custom_func))), silent = TRUE)
if (class(plt) == "try-error") {
  # replacement as it may return an error for some low-quality hashtags
  pairs(hash_prop, lower.panel = NULL,
      panel = function(...) smoothScatter(..., nrpoints = 0, add = TRUE))
} 

# plot the first hashtag of each sample only
ggplot(hash_prop, aes(x=hash_prop[,colnames(y)[y[1,] == 1][1]],
                       y=hash_prop[,colnames(y)[y[2,] == 1][1]])) +
   geom_point(alpha = 0.3, size = 0.5) +
    geom_density_2d(color = "orange") + 
  xlab(colnames(y)[y[1,] == 1][1]) + 
  ylab(colnames(y)[y[2,] == 1][1])



```


```{r}
sessionInfo()
```
