==> analysis_rmd/4_Ab_diagnostics.Rmd <==
---
title: "Ab diagnostics"
author: "Stephany Orjuela & Anthony Sonrel"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
    theme: yeti
    highlight: tango
    code_folding: show
    keep_md: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, dev = c("png", "pdf"))
execute <- TRUE
```

## Load packages


```{r}
suppressPackageStartupMessages({
 library(ggplot2)
 library(SingleCellExperiment)
 library(CATALYST)
 library(scater) 
 })
```

```{r}
inputFile <- "sce/d_rna_filt.rds" 
```

## Read debarcoded and filtered sce (after running 1, 2 and 3 steps)

These files have been filtered for:

* cells that could not be assigned to a sample_id.
* cross-sample doublets.
* within-sample doublets.
* RNA MAD5 per batch.
* ADT MAD5 per batch.

```{r}
d <- readRDS(inputFile)

d
altExp(d)
colData(altExp(d)) <- colData(d)[,1:6]

altExp(d) <- altExp(d)[rownames(altExp(d))!="MFRTK",]
 
```

## CLR normalize

Seurat-like normalization
```{r}
clr_function <- function(x) {
          return(log1p(x = x / (exp(x = sum(log1p(x = x[x > 0]), na.rm = TRUE) / length(x = x)))))
}

#by cells (2)
ab_norm <- apply(counts(altExp(d)), 2, clr_function)

#by proteins (1)
ab_norm_p <- t(apply(counts(altExp(d)), 1, clr_function))

assay(altExp(d), "CLRcounts") <- ab_norm
```

### Prepare object for use with CATALYST

```{r}

#add marker types
rowData(altExp(d))$marker_class <- "type"
rowData(altExp(d))$marker_class[grepl("IgG", rownames(altExp(d)))] <- "none"
rowData(altExp(d))$marker_name <- rowData(altExp(d))$ID

#turn coldata into factors
colData(altExp(d))$sample_id <- factor(colData(altExp(d))$sample_id)
colData(altExp(d))$patient_id <- factor(colData(altExp(d))$patient_id)
colData(altExp(d))$condition <- factor(colData(altExp(d))$condition)
colData(altExp(d))$Sample <- factor(colData(altExp(d))$Sample)
```

## Expression heatmap

```{r, fig.width = 10, fig.height= 10}
assay(altExp(d), "counts") <- as.matrix(assay(altExp(d), "counts"))

#Heatmap on raw counts, no scaling
plotExprHeatmap(altExp(d), features = "type", assay = "counts",
                bars = TRUE, row_clust = TRUE, scale = "never")

#Heatmap on raw counts, scaling
plotExprHeatmap(altExp(d), features = "type", assay = "counts",
                bars = FALSE, row_clust = TRUE)

#Heatmap on CLR, no scaling
plotExprHeatmap(altExp(d), features = "type", assay = "CLRcounts",
                bars = FALSE, row_clust = TRUE, scale = "never")

#Heatmap on CLR, scaling
plotExprHeatmap(altExp(d), features = "type", assay = "CLRcounts",
                bars = FALSE, row_clust = TRUE)

```

## Marker densities

```{r, fig.width = 10, fig.height= 10}
assay(altExp(d), "exprs") <- as.matrix(assay(altExp(d), "counts"))

#Smoothed densities of raw marker intensities, colored by well
plotExprs(altExp(d), color_by = "Sample") +  xlim(0, 50)


assay(altExp(d), "exprs") <- as.matrix(assay(altExp(d), "CLRcounts"))

#Smoothed densities of raCLR-normalized  marker intensities, colored by well
plotExprs(altExp(d), color_by = "Sample") +  xlim(0, 2)

#Smoothed densities of raCLR-normalized  marker intensities, colored by condition
plotExprs(altExp(d), color_by = "condition") +  xlim(0, 2)

```

```{r}

#num-features, sum(expr)
total_expr <- colSums(assay(altExp(d), "counts"))
num_features <- colSums(assay(altExp(d), "counts") > 0)
df <- data.frame(total_expr = total_expr,
                 num_features = num_features,
                 sample_id = colData(altExp(d))$sample_id,
                 well = colData(altExp(d))$Sample)

ggplot(df, aes(sample_id, log10(total_expr))) +
  geom_violin() 

ggplot(df, aes(well, log10(total_expr))) +
  geom_violin() 

ggplot(df, aes(sample_id,num_features)) +
  geom_violin() 
```


## MDS plot
```{r}
#pbMDS(altExp(d), color_by = "condition", assay = "counts")
pbMDS(altExp(d), color_by = "condition", assay = "CLRcounts")
pbMDS(altExp(d), color_by = "patient_id", assay = "CLRcounts")
pbMDS(altExp(d), color_by = "Sample", assay = "CLRcounts")
```

## Plot non-redundancy scores 
```{r, fig.width = 8, fig.height= 8}
assay(altExp(d), "exprs") <- as.matrix(assay(altExp(d), "CLRcounts"))
plotNRS(altExp(d))
```

## Dimension reduction
```{r}

set.seed(1010010)
altExp(d) <- runDR(altExp(d), cells = NULL, features = "type", assay = "CLRcounts")
plotDR(altExp(d)[,sample(1:ncol(d), ncol(d))], color_by="Sample") + geom_point(alpha = 0.3, size = 0.5)
plotDR(altExp(d)[,sample(1:ncol(d), ncol(d))], color_by="condition") + geom_point(alpha = 0.3, size = 0.5)
plotDR(altExp(d)[,sample(1:ncol(d), ncol(d))], color_by="patient_id") + geom_point(alpha = 0.3, size = 0.5)

```

```{r, fig.width = 10, fig.height= 10}
plotDR(altExp(d), color_by = type_markers(altExp(d)), assay = "CLRcounts", scale = FALSE) +
  geom_point(alpha = 0.3, size = 0.5)

plotDR(altExp(d), color_by = type_markers(altExp(d)), assay = "CLRcounts") +
  geom_point(alpha = 0.3, size = 0.5)

```


## Clustering

```{r}
library(scran)
set.seed(1010001)
g.adt <- buildSNNGraph(altExp(d), d=NA, assay.type = "CLRcounts", k = 8,
                       subset.row = type_markers(altExp(d)), 
                       BPPARAM = BiocParallel::MulticoreParam(5)) 

clusters.adt <- igraph::cluster_louvain(g.adt)
altExp(d)$cluster_id <- factor(clusters.adt$membership)

#Try cytof clustering
plotDR(altExp(d)[,sample(1:ncol(d), ncol(d))], color_by="cluster_id",facet_by = "condition") + geom_point(alpha = 0.3, size = 0.5)
```

## Cluster heatmap
```{r,fig.width = 10, fig.height= 10}
metadata(altExp(d))$cluster_codes <- data.frame(
  custom = factor(clusters.adt$membership), 
  levels = clusters.adt$membership)

plotExprHeatmap(altExp(d), by = "cluster_id", assay = "CLRcounts",
                k = "custom", features = "type", 
                scale = "never", bars = TRUE, row_clust = FALSE)

plotExprHeatmap(altExp(d), by = "cluster_id", assay = "CLRcounts",
                k = "custom", features = "type", 
                scale = "first", bars = TRUE, row_clust = FALSE)

saveRDS(d,"sce/d_adt_clust.rds")
```

```{r}
sessionInfo()
```
